Refactor pixel_loader.v to simplify addr reset
The code in pixel_loader.v has been refactored to simplify the initialization process. Instead of using individual if statements for each color and address, a single block of code has been added to set all addresses to 0 at the beginning of each frame. This change improves readability and reduces redundancy in the code.